# packages/types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backline API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/revisions": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "List Project Revisions",
        "operationId": "list_project_revisions_api_v1_projects__project_id__revisions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RevisionHistoryOut"
                  },
                  "title": "Response List Project Revisions Api V1 Projects  Project Id  Revisions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pages/{page_id}/snapshots": {
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Submit Snapshot",
        "operationId": "submit_snapshot_api_v1_pages__page_id__snapshots_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Page Id"
            }
          },
          {
            "name": "x-guest-session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Guest-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotSubmit"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RevisionChangeSummary": {
        "properties": {
          "moved": {
            "type": "integer",
            "title": "Moved",
            "default": 0
          },
          "modified": {
            "type": "integer",
            "title": "Modified",
            "default": 0
          },
          "removed": {
            "type": "integer",
            "title": "Removed",
            "default": 0
          },
          "added": {
            "type": "integer",
            "title": "Added",
            "default": 0
          }
        },
        "type": "object",
        "title": "RevisionChangeSummary"
      },
      "RevisionHistoryOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "page_id": {
            "type": "string",
            "title": "Page Id"
          },
          "page_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Title"
          },
          "page_url": {
            "type": "string",
            "title": "Page Url"
          },
          "full_page_hash": {
            "type": "string",
            "title": "Full Page Hash"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current"
          },
          "changes": {
            "$ref": "#/components/schemas/RevisionChangeSummary"
          },
          "recovery": {
            "$ref": "#/components/schemas/RevisionRecoverySummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "page_id",
          "page_title",
          "page_url",
          "full_page_hash",
          "captured_at",
          "is_current",
          "changes",
          "recovery"
        ],
        "title": "RevisionHistoryOut"
      },
      "RevisionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "page_id": {
            "type": "string",
            "title": "Page Id"
          },
          "full_page_hash": {
            "type": "string",
            "title": "Full Page Hash"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time",
            "title": "Captured At"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current"
          },
          "created_new": {
            "type": "boolean",
            "title": "Created New"
          }
        },
        "type": "object",
        "required": [
          "id",
          "page_id",
          "full_page_hash",
          "captured_at",
          "is_current",
          "created_new"
        ],
        "title": "RevisionOut"
      },
      "RevisionRecoverySummary": {
        "properties": {
          "ok": {
            "type": "integer",
            "title": "Ok",
            "default": 0
          },
          "low_confidence": {
            "type": "integer",
            "title": "Low Confidence",
            "default": 0
          },
          "orphaned": {
            "type": "integer",
            "title": "Orphaned",
            "default": 0
          },
          "permanently_orphaned": {
            "type": "integer",
            "title": "Permanently Orphaned",
            "default": 0
          }
        },
        "type": "object",
        "title": "RevisionRecoverySummary"
      },
...
      "SnapshotSubmit": {
        "properties": {
          "viewport": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Viewport"
          },
          "node_tree": {
            "type": "object",
            "title": "Node Tree"
          },
          "nodes_index": {
            "type": "object",
            "title": "Nodes Index"
          },
          "full_page_hash": {
            "type": "string",
            "title": "Full Page Hash"
          }
        },
        "type": "object",
        "required": [
          "viewport",
          "node_tree",
          "nodes_index",
          "full_page_hash"
        ],
        "title": "SnapshotSubmit"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}